pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nubi"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "kopf",
    "pydantic>=2.0",
    "kubernetes-asyncio",
    "strands-agents",
    "aiohttp",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nubi"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["kubernetes_asyncio.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["strands.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["aiohttp.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
